refactor(examples): rearrange server functionality logic
