+++ Examples/src/server/routes/variants.ts
